+++ Kindy/Kindy/View Controllers/Home/HomeViewController+CompositionalLayout.swift
@@ -1,0 +1,163 @@
+//
+//  HomeViewController+CompositionalLayout.swift
+//  Kindy
+//
+//  Created by 정호윤 on 2022/12/03.
+//
+
+import UIKit
+
+extension HomeViewController {    
+    func createLayout() -> UICollectionViewLayout {
+        let layout = UICollectionViewCompositionalLayout { sectionIndex, layoutEnvironment in
+            let section = self.dataSource.snapshot().sectionIdentifiers[sectionIndex]
+            
+            // MARK: Section Header
+            let headerItemSize = NSCollectionLayoutSize(
+                widthDimension: .fractionalWidth(1),
+                heightDimension: .estimated(36)
+            )
+            let headerItem = NSCollectionLayoutBoundarySupplementaryItem(
+                layoutSize: headerItemSize,
+                elementKind: SupplementaryViewKind.header,
+                alignment: .top
+            )
+            headerItem.contentInsets = NSDirectionalEdgeInsets(top: .zero, leading: padding16, bottom: .zero, trailing: padding16)
+            
+            // MARK: Item Size
+            let fullSize = NSCollectionLayoutSize(
+                widthDimension: .fractionalWidth(1),
+                heightDimension: .fractionalHeight(1)
+            )
+            
+            // MARK: Content Insets
+            let leading16ContentInsetsForItem = NSDirectionalEdgeInsets(top: .zero, leading: padding16, bottom: .zero, trailing: .zero)
+            let sectionContentInsets = NSDirectionalEdgeInsets(top: padding8, leading: .zero, bottom: padding8, trailing: .zero)
+            
+            switch section {
+            case .curations:
+                let item = NSCollectionLayoutItem(layoutSize: fullSize)
+                
+                let groupSize = NSCollectionLayoutSize(
+                    widthDimension: .fractionalWidth(0.92),
+                    heightDimension: .estimated(408)
+                )
+                let group = NSCollectionLayoutGroup.horizontal(
+                    layoutSize: groupSize,
+                    subitem: item,
+                    count: 1
+                )
+                
+                let section = NSCollectionLayoutSection(group: group)
+                section.orthogonalScrollingBehavior = .groupPagingCentered
+                section.boundarySupplementaryItems = [headerItem]
+                section.contentInsets = sectionContentInsets
+                
+                return section
+            case .featured:
+                let item = NSCollectionLayoutItem(layoutSize: fullSize)
+                item.contentInsets = leading16ContentInsetsForItem
+                
+                let groupSize = NSCollectionLayoutSize(
+                    widthDimension: .estimated(326),
+                    heightDimension: .estimated(168)
+                )
+                let group = NSCollectionLayoutGroup.horizontal(
+                    layoutSize: groupSize,
+                    subitems: [item]
+                )
+                
+                let section = NSCollectionLayoutSection(group: group)
+                section.orthogonalScrollingBehavior = .groupPaging
+                section.boundarySupplementaryItems = [headerItem]
+                section.contentInsets = sectionContentInsets
+                
+                return section
+            case .nearbys:
+                let itemSize = NSCollectionLayoutSize(
+                    widthDimension: .fractionalWidth(1),
+                    heightDimension: .fractionalHeight(1/3)
+                )
+                let item = NSCollectionLayoutItem(layoutSize: itemSize)
+                
+                let groupSize = NSCollectionLayoutSize(
+                    widthDimension: .fractionalWidth(0.92),
+                    heightDimension: .estimated(312)
+                )
+                let group = NSCollectionLayoutGroup.vertical(
+                    layoutSize: groupSize,
+                    subitem: item,
+                    count: 3
+                )
+                
+                let section = NSCollectionLayoutSection(group: group)
+                section.orthogonalScrollingBehavior = .groupPagingCentered
+                section.boundarySupplementaryItems = [headerItem]
+                section.contentInsets = sectionContentInsets
+                
+                return section
+            case .bookmarks:
+                let item = NSCollectionLayoutItem(layoutSize: fullSize)
+                item.contentInsets = leading16ContentInsetsForItem
+                
+                let groupSize = NSCollectionLayoutSize(
+                    widthDimension: .fractionalWidth(2/5),
+                    heightDimension: .estimated(219)
+                )
+                let group = NSCollectionLayoutGroup.horizontal(
+                    layoutSize: groupSize,
+                    subitems: [item]
+                )
+                
+                let section = NSCollectionLayoutSection(group: group)
+                section.orthogonalScrollingBehavior = .continuousGroupLeadingBoundary
+                section.boundarySupplementaryItems = [headerItem]
+                section.contentInsets = sectionContentInsets
+                
+                return section
+            case .regions:
+                let itemSize = NSCollectionLayoutSize(
+                    widthDimension: .fractionalWidth(1/2),
+                    heightDimension: .fractionalHeight(1)
+                )
+                let item = NSCollectionLayoutItem(layoutSize: itemSize)
+                
+                let groupSize = NSCollectionLayoutSize(
+                    widthDimension: .fractionalWidth(1),
+                    heightDimension: .estimated(50)
+                )
+                let group = NSCollectionLayoutGroup.horizontal(
+                    layoutSize: groupSize,
+                    subitem: item,
+                    count: 2
+                )
+                
+                let section = NSCollectionLayoutSection(group: group)
+                section.boundarySupplementaryItems = [headerItem]
+                section.contentInsets = sectionContentInsets
+                
+                return section
+            case .noPermission, .emptyNearbys, .emptyBookmarks:
+                let item = NSCollectionLayoutItem(layoutSize: fullSize)
+                
+                let groupSize = NSCollectionLayoutSize(
+                    widthDimension: .fractionalWidth(1),
+                    heightDimension: .estimated(150)
+                )
+                let group = NSCollectionLayoutGroup.horizontal(
+                    layoutSize: groupSize,
+                    subitem: item,
+                    count: 1
+                )
+                
+                let section = NSCollectionLayoutSection(group: group)
+                section.boundarySupplementaryItems = [headerItem]
+                section.contentInsets = sectionContentInsets
+                
+                return section
+            }
+        }
+        
+        return layout
+    }
+}
